fix(parquet): FSST review — bound the worst-case payload, and check what actually decompressed
Three of Copilot's four findings on #149, plus tests for the two behavioural ones.

TryCompress summed the chunk's raw bytes into an int and doubled it for §5.3's worst-case
payload bound. The sum itself cannot overflow — Arrow's int32 offsets cap a BinaryArray's
total — but the DOUBLING wraps negative somewhere past a 1 GB chunk, turning a size question
into an allocation crash. Now accumulated and doubled in long, with a chunk whose worst case
will not fit an array simply declined: the caller already falls back for "FSST did not shrink
this", and this is the same answer. MaxCompressedLength takes and returns long for the same
reason, so it cannot silently wrap at a future call site.

ReadSymbolTablePage ignored the byte count Decompressor.Decompress returns and sliced the
buffer to the page's DECLARED uncompressed size. Those differ only on a corrupt page — but
the buffer is rented from ArrayPool, so it arrives holding whatever the last renter left in
it, and a short decompression would hand Parse stale bytes it could conceivably validate as
a well-formed symbol table. That is exactly the silent-wrong-data case §3.7 asks readers to
prevent, and the symbol table is the worst place to have it since every data page in the
chunk decodes against it. Now checked, along with a negative declared size, which previously
surfaced as ArgumentOutOfRangeException from the slice rather than a format error.

Not changed: the same review flagged that a page whose is_compressed flag is set under an
Uncompressed codec is parsed as plain. It is contradictory, but there is no codec to invoke
and the bytes can only be plain — erroring would reject a file that reads correctly. Now
commented at the branch, and pinned by a test so the behaviour is deliberate rather than
incidental.

Docs said `SymbolTableType.FSST`, which is the spec's spelling of the variant wearing the C#
type's name; the enum member is `SymbolTableType.Fsst`. Both sites now name the spec variant
plainly, with the C# spelling given once where the API is what is meant.

841 green on net10.0 and net8.0, 839 on net472.

+++ src/EngineeredWood.Parquet/Parquet/Data/FsstSymbolTable.cs
@@ -239,10 +239,14 @@ internal sealed class FsstSymbolTable
     }
 
     /// <summary>
-    /// Upper bound on the compressed size of a <paramref name="rawLength"/>-byte value:
+    /// Upper bound on the compressed size of a <paramref name="rawLength"/>-byte input:
     /// every byte escaping to a two-byte sequence (§5.3).
     /// </summary>
-    public static int MaxCompressedLength(int rawLength) => rawLength * 2;
+    /// <remarks>
+    /// Takes and returns <see cref="long"/> so the doubling cannot silently wrap for a
+    /// column chunk over 1 GB — callers bound the result before allocating.
+    /// </remarks>
+    public static long MaxCompressedLength(long rawLength) => rawLength * 2;
 
     /// <summary>
     /// Compresses one value into <paramref name="destination"/> and returns the byte count.
